+++ CHANGELOG.md
@@ -20,6 +20,7 @@
     * Check in `Settings` that `psycopg2`, `asyngpg` and `cfut`, if required, are installed (\#1167).
     * Split `DB.set_db` into sync/async methods (\#1165).
     * Rename `DB.get_db` into `DB.get_async_db` (\#1183).
+    * Normalize names of task packages (\#1188).
 * Testing:
     * Update `clean_db_fractal_1.4.1.sql` to `clean_db_fractal_1.4.2.sql`, and change `migrations.yml` target version (\#1152).
     * Reorganise the test directory into subdirectories, named according to the order in which we want the CI to execute them (\#1166).
@@ -29,6 +30,7 @@
     * Remove `JobStatusType.RUNNING`, incorporating it into `JobStatusType.SUBMITTED` (\#1179).
 * Benchmarks:
     * Add `fractal_client.py` and `populate_script_v2.py` for creating different database status scenarios.
+
 # 1.4.2
 
 > **WARNINGs**: